requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
pdfplumber==0.7.6
fpdf==1.7.2
torch==2.0.1
transformers==4.30.2
nltk==3.8.1
arabic-reshaper==2.1.0
python-bidi==0.4.2
google-generativeai==0.0.6
unicodedata2==14.0.0
